Optimized the way the X-Axis labels are displayed.
The graph will now offset the labels on the X-Axis if needed and
possible in order for them to be centered.

A good example is with 12 points and a gap of 2:
- Before, the first label would be displayed on the 3rd point and the
last label was displayed on the last point.
- Now, all of the labels are shifted one row to the left so that the
first label is displayed on the second point and the last label on the
second to last point. This way, the labels appear centered on the
X-Axis.