Fixed Refernce Line Alpha Bug
Fixed an issue where the reference lines would have an alpha value of 0
if the line also had an alpha value of zero. Reference lines now set to
an alpha of 0.1 when the line alpha is 0.0. As before, you can disable
reference lines using the boolean properties to make them appear or
disappear.