New Average Line; New Animation
Added an Average Line feature. Draw an average line with a specific
y-value. Use the new averageLine property on BEMSimpleLineGraph to
setup and customize the line. Might be considered a fix for issue #42.
The implementation of the average line feature is likely the direction
BEMSimpleLineGraph is headed as it expands.

A new “expansion” animation has been added to the list of available
animations. Try out the new animation with the BEMLineAnimationExpand
type.